feat: Combine imageLoadReducers
